voice_transcriber/requirements.txt
flask
openai-whisper